pyproject.toml
[tool.poetry]
name = "dev-bot"
version = "0.1.0"
description = "Slack bot to increase productivity"
authors = ["CTDS UChicago <cdis@uchicago.edu>"]

[tool.poetry.dependencies]
python = "^3.6"
google-api-python-client = "^2.15.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.4.4"
aiohttp = "^3.7.4"
slackeventsapi = "^2.2.1"
slackclient = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
